(PC-31390)[API] feat: rename folder `simulate_adage_steps` to `adage_mock`
