+++ api/src/pcapi/routes/public/collective/endpoints/adage_mock/errors.py
